requirements.txt
streamlit==1.31.0
supabase==2.3.0
python-dotenv==1.0.0
plotly==5.18.0
pandas==2.1.4
anthropic==0.18.1
numpy==1.26.3
python-dateutil==2.8.2
